Changed datatype of id of Entity from int to long.

+++ Source/ch/nolix/core/databaseAdapter/EntitySet.java
@@ -166,7 +166,7 @@ public final class EntitySet<E extends Entity> extends NamedElement {
 	
 	//method
 	@SuppressWarnings("unchecked")
-	public E getRefEntityById(final int id) {
+	public E getRefEntityById(final long id) {
 		
 		final var loadedEntity = loadedAndCreatedEntities.getRefFirstOrNull(e -> e.getId() == id);
 		if (loadedEntity != null) {
